requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.340
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
pypdf==3.17.1
python-docx==1.1.0
openai==1.3.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
torch==2.1.0
transformers==4.36.0
numpy==1.24.3
jsonlines